fix(sigv4): Allow custom `scheme`
Forwards custom schemes through to signed URLs
